requirements.txt
Flask==3.0.3
qrcode[pil]==8.0
Pillow==10.3.0
